Add a store test
