+++ src/__TESTS__/store.test.js
@@ -1,0 +1,9 @@
+import '@testing-library/jest-dom';
+import store from '../redux/store';
+
+describe('The store object', () => {
+  it('should validate redux store object', () => {
+    expect(store).toBeTruthy();
+    expect(typeof store.dispatch).toBe('function');
+  });
+});
